# openapi-fastapi.yaml
openapi: 3.0.3
info:
  title: SKAX Physical Risk AI Agent API (FastAPI)
  version: 1.0.0
  description: >
    사업장 기후 물리적 리스크 AI Agent 분석 시스템 API

    Spring Boot 서버로부터 요청을 받아 AI Agent를 통해 기후 리스크를 분석하고 결과를 반환합니다.

    주요 기능:
    - AI Agent 기반 물리적 리스크 분석
    - 비동기 작업 처리 및 상태 관리
    - 과거 재난 이력 분석
    - SSP 시나리오별 미래 리스크 전망
    - 재무 영향 분석 (AAL, CVaR)
    - 취약성 평가
    - 사업장 이전 시뮬레이션
    - 기후 변화 시뮬레이션
servers:
  - url: https://ai-api.skax.co.kr
    description: Production
  - url: https://staging-ai-api.skax.co.kr
    description: Staging
  - url: http://localhost:8000
    description: Local Development
...
components:
  securitySchemes:
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: Spring Boot 서버 인증용 API Key
  schemas:
...
    SiteInfo:
      type: object
      required: [id, name, latitude, longitude, industry]
      properties:
        id:
          type: string
          format: uuid
          description: 사업장 ID
        name:
          type: string
          example: 대덕 데이터센터
        address:
          type: string
          example: 대전광역시 유성구 대덕대로 512
        city:
          type: string
          example: 대전광역시
        latitude:
          type: number
          format: double
          example: 36.368
        longitude:
          type: number
          format: double
          example: 127.364
        industry:
          type: string
          example: DC(Data Center)
          description: 업종
        buildingAge:
          type: integer
          description: 건물 연령 (년)
          example: 15
        buildingType:
          type: string
          description: 건물 유형
          example: 철근 콘크리트
        seismicDesign:
          type: boolean
          description: 내진설계 여부
          default: false
        floorArea:
          type: number
          format: double
          description: 연면적 (m²)
        assetValue:
          type: number
          format: double
          description: 자산 가치 (KRW)
        employeeCount:
          type: integer
          description: 직원 수

    # ==========================================
    # Analysis Request/Response Schemas
    # ==========================================
...
    RelocationCandidate:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          example: 평택 산업단지
        city:
          type: string
          example: 경기도 평택시
        latitude:
          type: number
          format: double
        longitude:
          type: number
          format: double
        locationSummary:
          type: string
          example: 경기도 평택시, 산업단지 인근
        riskScore:
          type: integer
          minimum: 0
          maximum: 100
          description: 종합 리스크 점수
        riskLevel:
          type: string
          enum: [낮음, 보통, 높음, 심각]
        aal:
          type: number
          format: double
          description: Average Annual Loss (KRW)
        distance:
          type: number
          format: float
          description: 현재 사업장으로부터 거리 (km)
        estimatedCost:
          type: number
          format: double
          description: 예상 이전 비용 (KRW)
        advantages:
          type: array
          items:
            type: string
          example:
            - 홍수 리스크 낮음
            - 태풍 영향 적음
        disadvantages:
          type: array
          items:
            type: string
          example:
            - 폭염 리스크 소폭 증가
    RelocationCandidatesRequest:
      type: object
      required: [baseSiteId]
      properties:
        baseSiteId:
          type: string
          format: uuid
          description: 기준 사업장 ID (Agent가 자동으로 최적 후보지를 추천)
    RelocationSimulationRequest:
      type: object
      required: [baseSiteId, candidateLatitude, candidateLongitude]
      properties:
        baseSiteId:
          type: string
          format: uuid
          description: 기준 사업장 ID
        candidateLatitude:
          type: number
          format: double
          description: 후보지 위도
          example: 37.0892
        candidateLongitude:
          type: number
          format: double
          description: 후보지 경도
          example: 127.0548
        candidateName:
          type: string
          description: 후보지 이름 (선택사항)
          example: 평택 산업단지
    RelocationSimulationResult:
      type: object
      properties:
        baseSite:
          type: object
          properties:
            id:
              type: string
            name:
              type: string
            riskScore:
              type: integer
            aal:
              type: number
              format: float
              description: 연평균 자산 손실률 (AAL)
            annualExpectedLoss:
              type: number
              format: double
              description: 연평균 재무 손실액 (KRW)
        candidateSite:
          type: object
          properties:
            name:
              type: string
            latitude:
              type: number
              format: double
            longitude:
              type: number
              format: double
            riskScore:
              type: integer
            aal:
              type: number
              format: float
              description: 연평균 자산 손실률 (AAL)
            annualExpectedLoss:
              type: number
              format: double
              description: 연평균 재무 손실액 (KRW)
        comparison:
          type: object
          properties:
            riskScoreDiff:
              type: integer
              description: 리스크 점수 차이 (음수면 개선)
            riskScoreDiffPercent:
              type: number
              format: float
              description: 리스크 점수 차이 비율 (%)
            aalDiff:
              type: number
              format: float
              description: AAL 차이 (음수면 개선)
            annualExpectedLossDiff:
              type: number
              format: double
              description: 연평균 재무 손실액 차이 (KRW, 음수면 절감)
        climateFactorComparison:
          type: object
          description: 기후 요인별 비교
          additionalProperties:
            type: object
            properties:
              hazardType:
                type: string
              base:
                type: integer
                description: 현재 사업장 점수
              candidate:
                type: integer
                description: 후보지 점수
              diff:
                type: integer
                description: 점수 차이
              diffPercent:
                type: number
                format: float
        pros:
          type: array
          description: 이전 시 장점
          items:
            type: string
          example:
            - 홍수 리스크 30% 감소
            - 태풍 빈도 낮음
            - 연간 예상 손실액 20억원 감소
        cons:
          type: array
          description: 이전 시 단점
          items:
            type: string
          example:
            - 폭염 리스크 10% 증가
            - 이전 비용 약 50억원
        recommendation:
          type: string
          description: AI가 생성한 추천 의견
          example: 장기적 리스크 감소를 고려할 때 이전을 권장합니다.
        financialAnalysis:
          type: object
          properties:
            relocationCost:
              type: number
              format: double
            annualSavings:
              type: number
              format: double
              description: 연간 절감 예상액
            breakEvenYears:
              type: number
              format: float
              description: 손익분기점 (년)
            npv10Years:
              type: number
              format: double
              description: 10년 순현재가치
    ClimateSimulationRequest:
      type: object
      required: [scenario, sites]
      properties:
        scenario:
          type: string
          enum: [SSP1-2.6, SSP2-4.5, SSP3-7.0, SSP5-8.5]
          example: SSP5-8.5
        hazardType:
          type: string
          example: 폭염
          description: 특정 Hazard 지정 (미지정시 전체)
        sites:
          type: array
          description: 비교할 사업장 목록
          items:
            $ref: '#/components/schemas/SiteInfo'
        startYear:
          type: integer
          example: 2025
          default: 2025
        endYear:
          type: integer
          example: 2100
          default: 2100
        timeStep:
          type: integer
          description: 시간 간격 (년)
          default: 10
    ClimateSimulationPoint:
      type: object
      properties:
        year:
          type: integer
          example: 2045
        metric:
          type: number
          format: float
          description: 측정값 (온도, 강수량 등)
        riskScore:
          type: integer
          description: 해당 시점의 리스크 점수
    ClimateSimulationSeries:
      type: object
      properties:
        siteId:
          type: string
          format: uuid
        siteName:
          type: string
        metricName:
          type: string
          example: avg_temperature
          description: 측정 지표명
        metricUnit:
          type: string
          example: "°C"
        points:
          type: array
          items:
            $ref: '#/components/schemas/ClimateSimulationPoint'
    ClimateSimulationResponse:
      type: object
      properties:
        scenario:
          type: string
          example: SSP5-8.5
        scenarioDescription:
          type: string
        hazardType:
          type: string
        series:
          type: array
          description: 사업장별 시계열 데이터
          items:
            $ref: '#/components/schemas/ClimateSimulationSeries'
        timelineActions:
          type: array
          description: 시간대별 권장 조치
          items:
            type: object
            properties:
              decade:
                type: string
                example: "2030s"
              year:
                type: integer
                example: 2035
              action:
                type: string
                example: 냉각 시스템 업그레이드
              targetSites:
                type: array
                description: 해당 조치가 필요한 사업장
                items:
                  type: string
              priority:
                type: string
                enum: [low, medium, high, critical]
              estimatedCost:
                type: number
                format: double
              expectedBenefit:
                type: string
        summary:
          type: string
          description: AI가 생성한 시뮬레이션 요약

    # ==========================================
    # Report Schemas
    # ==========================================
...
paths:
...
  /api/v1/simulation/relocation/candidates:
    post:
      tags: [Simulation]
      summary: 사업장 이전 후보지 추천
      description: >
        AI Agent가 현재 사업장 대비 리스크가 낮은 후보지를 추천합니다.
        각 후보지에 대해 리스크 분석을 수행합니다.
      security:
        - apiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RelocationCandidatesRequest'
      responses:
        '200':
          description: 후보지 목록
          content:
            application/json:
              schema:
                type: object
                properties:
                  baseSite:
                    type: object
                    properties:
                      id:
                        type: string
                      name:
                        type: string
                      riskScore:
                        type: integer
                  candidates:
                    type: array
                    items:
                      $ref: '#/components/schemas/RelocationCandidate'
        '400':
          description: 잘못된 요청
  /api/v1/simulation/relocation/compare:
    post:
      tags: [Simulation]
      summary: 사업장 이전 시뮬레이션 (비교)
      description: >
        현재 사업장과 후보지의 리스크를 상세 비교합니다.
        두 사업장에 대해 전체 분석을 수행하고 비교 결과를 반환합니다.
      security:
        - apiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RelocationSimulationRequest'
      responses:
        '200':
          description: 비교 결과
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RelocationSimulationResult'
        '400':
          description: 잘못된 요청
  /api/v1/simulation/climate:
    post:
      tags: [Simulation]
      summary: 기후 시뮬레이션
      description: >
        SSP 시나리오에 따른 미래 기후 변화를 시뮬레이션합니다.
        여러 사업장을 동시에 비교할 수 있습니다.
      security:
        - apiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClimateSimulationRequest'
      responses:
        '200':
          description: 시뮬레이션 결과
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClimateSimulationResponse'
        '400':
          description: 잘못된 요청

  # ==========================================
  # Reports Endpoints
  # ==========================================
...
security:
  - apiKeyAuth: []